change timing push notification alert
