+++ Ani/View/OtherProfile/ANIOtherProfileBasicView.swift
@@ -585,7 +585,7 @@ extension ANIOtherProfileBasicView: UITableViewDelegate {
       cell.unobserveLove()
       cell.unobserveSupport()
     } else if contentType == .story {
-      if !stories.isEmpty {
+      if !stories.isEmpty && stories.count > indexPath.row {
         if stories[indexPath.row].recruitId != nil, let cell = cell as? ANISupportViewCell {
           cell.unobserveLove()
           cell.unobserveComment()
